+++ spys/WriteSpy.h
@@ -19,7 +19,7 @@ void WriteSpy_Write32( uint32_t cmd );
 
 void WriteSpy_Write8( uint8_t cmd );
 
-void WriteSpy_Write8Arr( uint8_t* cmd , uint32_t size );
+void WriteSpy_Write8Arr( uint8_t* cmd , uint16_t size );
 
 #endif
